Pause Screen mostly finished
Click on red block in upper right corner of screen to pause the game. A
pause screen is drawn with three buttons. The top resumes the game, the
middle is a toggle for the sound, and bottom quits the game. The game
can also be paused and resumed with the 'p' key.