Placed animal loading into separate file
Created a new file called animals, meant to contain the code responsible
for loading animals and their respective attributes.  Most, if not all,
of the code from the animals team can be found in animals.lua.
NOTE: The additional file means that you must keep "require 'animals'"
at the top of main.